+++ src/matte.rs
@@ -4,7 +4,7 @@
 // Copyright (c) 2019-2020  Jeron Aldaron Lau
 //
 //! Matte color model and types.
-use crate::chan::{Ch16, Ch32, Ch8, Channel, Linear, Straight};
+use crate::chan::{Ch16, Ch32, Ch8, Channel, Linear, Premultiplied};
 use crate::el::{Pix1, PixRgba, Pixel};
 use crate::ColorModel;
 use std::ops::Range;
@@ -42,17 +42,14 @@ impl ColorModel for Matte {
     }
 }
 
-/// [Matte](struct.Matte.html) 8-bit [straight](../chan/struct.Straight.html)
-/// alpha [linear](../chan/struct.Linear.html) gamma
-/// [pixel](../el/trait.Pixel.html) format.
-pub type Matte8 = Pix1<Ch8, Matte, Straight, Linear>;
+/// [Matte](struct.Matte.html) 8-bit alpha [linear](../chan/struct.Linear.html)
+/// gamma [pixel](../el/trait.Pixel.html) format.
+pub type Matte8 = Pix1<Ch8, Matte, Premultiplied, Linear>;
 
-/// [Matte](struct.Matte.html) 16-bit [straight](../chan/struct.Straight.html)
-/// alpha [linear](../chan/struct.Linear.html) gamma
-/// [pixel](../el/trait.Pixel.html) format.
-pub type Matte16 = Pix1<Ch16, Matte, Straight, Linear>;
+/// [Matte](struct.Matte.html) 16-bit alpha [linear](../chan/struct.Linear.html)
+/// gamma [pixel](../el/trait.Pixel.html) format.
+pub type Matte16 = Pix1<Ch16, Matte, Premultiplied, Linear>;
 
-/// [Matte](struct.Matte.html) 32-bit [straight](../chan/struct.Straight.html)
-/// alpha [linear](../chan/struct.Linear.html) gamma
-/// [pixel](../el/trait.Pixel.html) format.
-pub type Matte32 = Pix1<Ch32, Matte, Straight, Linear>;
+/// [Matte](struct.Matte.html) 32-bit alpha [linear](../chan/struct.Linear.html)
+/// gamma [pixel](../el/trait.Pixel.html) format.
+pub type Matte32 = Pix1<Ch32, Matte, Premultiplied, Linear>;