+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+## [0.13.0] - 2020-05-05
 ### Changed
 * `Raster::rows()` / `rows_mut()` sliced to span of region
 * Renamed `Pixel::composite_channels_matte` to `composite_channels_alpha`